// src/domain/types/MappedPalette.ts
export type MappedPalette = {
  text: string;
  textSecondary: string;
  textTertiary: string;
  textInverse: string;

  background: string;
  surface: string;
  surfaceVariant: string;
  card: string;

  icon: string;
  tabIconDefault: string;
  tabIconSelected: string;

  primary: string;
  secondary: string;
  tertiary: string;

  border: string;
  borderVariant: string;

  success: string;
  successLight: string;
  warning: string;
  warningLight: string;
  error: string;
  errorLight: string;
  info: string;
  infoLight: string;

  button: string;
  buttonSecondary: string;

  fontPrimary: string;
  fontSecondary: string;
  fontTertiary: string;
};